Commit Final before Deployment
The aim of this project is to design, develop and evaluate a website. The intended audience of the website is the Test Valley Men’s Shed community for members to share projects, communicate and create user accounts. This involves requirements gathering, project management, research and evaluation of various platforms and technologies to determine the most appropriate method to fulfil the objectives of the project.